extract the numerous components inside UserProfile
